requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
google-genai>=0.2.0
websockets>=12.0
python-dotenv>=1.0.0
pydantic>=2.6.0
pydantic-settings>=2.1.0